Migrate project from React to Next.js with API routes and caching
