+++ app/globals.css
@@ -1,5 +1,3 @@
-@import url("https://fonts.googleapis.com/css2?family=JetBrains+Mono:ital,wght@0,100..800;1,100..800&display=swap");
-
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
@@ -71,8 +69,8 @@
 }
 
 :root {
-	font-family: "JetBrains Mono", monospace, sans-serif, Inter, system-ui,
-		Avenir, Helvetica, Arial, sans-serif;
+	font-family: monospace, sans-serif, Inter, system-ui, Avenir, Helvetica,
+		Arial, sans-serif;
 	font-optical-sizing: auto;
 	font-style: normal;
 	line-height: 1.5;
