+++ app/layout.tsx
@@ -1,0 +1,32 @@
+import type { Metadata } from "next";
+import { JetBrains_Mono } from "next/font/google";
+import "./globals.css";
+import { ThemeProvider } from "components/theme-provider";
+
+const jetbrains_mono = JetBrains_Mono({ subsets: ["latin"] });
+
+export const metadata: Metadata = {
+	title: "Dropbox",
+	description: "Dropbox clone",
+};
+
+export default function RootLayout({
+	children,
+}: Readonly<{
+	children: React.ReactNode;
+}>) {
+	return (
+		<html lang='en' suppressHydrationWarning={true}>
+			<body className={jetbrains_mono.className}>
+				<ThemeProvider
+					attribute='class'
+					defaultTheme='system'
+					enableSystem
+					disableTransitionOnChange
+				>
+					{children}
+				</ThemeProvider>
+			</body>
+		</html>
+	);
+}
