+++ components.json
@@ -1,11 +1,11 @@
 {
 	"$schema": "https://ui.shadcn.com/schema.json",
 	"style": "default",
-	"rsc": false,
+	"rsc": true,
 	"tsx": true,
 	"tailwind": {
-		"config": "tailwind.config.js",
-		"css": "src/index.css",
+		"config": "tailwind.config.ts",
+		"css": "app/globals.css",
 		"baseColor": "slate",
 		"cssVariables": true,
 		"prefix": ""
