+++ package.json
@@ -4,11 +4,10 @@
 	"version": "0.0.0",
 	"type": "module",
 	"scripts": {
-		"dev": "vite",
-		"dev-host": "vite --host",
-		"build": "tsc -b && vite build",
-		"lint": "eslint .",
-		"preview": "vite preview"
+		"dev": "next dev",
+		"start": "next start",
+		"build": "next build",
+		"lint": "eslint ."
 	},
 	"dependencies": {
 		"@radix-ui/react-dialog": "^1.1.1",
@@ -22,6 +21,8 @@
 		"framer-motion": "^11.3.24",
 		"lucide-react": "^0.426.0",
 		"mini-svg-data-uri": "^1.4.4",
+		"next": "^14.2.11",
+		"next-themes": "^0.3.0",
 		"react": "^18.3.1",
 		"react-day-picker": "8.10.1",
 		"react-dom": "^18.3.1",
@@ -34,7 +35,6 @@
 		"@types/node": "^22.1.0",
 		"@types/react": "^18.3.3",
 		"@types/react-dom": "^18.3.0",
-		"@vitejs/plugin-react-swc": "^3.5.0",
 		"autoprefixer": "^10.4.20",
 		"eslint": "^9.8.0",
 		"eslint-plugin-react-hooks": "^5.1.0-rc.0",
@@ -43,7 +43,6 @@
 		"postcss": "^8.4.41",
 		"tailwindcss": "^3.4.9",
 		"typescript": "^5.5.3",
-		"typescript-eslint": "^8.0.0",
-		"vite": "^5.4.0"
+		"typescript-eslint": "^8.0.0"
 	}
 }
