+++ typings.ts
@@ -1,0 +1,5 @@
+export interface Holiday {
+	name: string;
+	dates: string[];
+	description: string;
+}
